Centralize all pass/chunk related code into reader_impl_chunking.cu
